Add files via upload
In order to prevent the spread of CORONA virus, everyone must wear a mask. In these tough times of COVID-19, it is necessary to build a model that detects people with and without mask in real-time. As it works as a simple precautionary measure to prevent the spread of virus